+++ src/components/Magazine/index.jsx
@@ -1,6 +1,7 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import Header from '../layout/Header'
 import Main from './Main'
+import GoTop from '../common/GoTop'
 
 const Magazine = () => {
   return (
@@ -9,6 +10,7 @@ const Magazine = () => {
       <div className="pt-[54px] pb-[66px]">
         <Main />
       </div>
+      <GoTop />
     </>
   )
 }
